test: move rate_from_bracket_indice test to test_rate_tax_scale_like
